requirements.txt
fastapi>=0.104.1
pydantic>=2.4.2
uvicorn>=0.24.0.post1
sqlalchemy>=2.0.23
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
python-multipart>=0.0.6
aiofiles>=23.2.1
requests>=2.31.0
pytest>=7.4.3
httpx>=0.25.1
python-dotenv>=1.0.0